Messing with error handling + wrappers
Attempting to utilize wrappers to avoid the try/except/finally blocks
being all over the place. However, it seems like there are some
wonks with passing the session around. For the time being it looks
like just keeping the transactions in the same location as the
session to be the only way of doing things.